# _15_appliance_store/api/store/v1/financialTransaction_router.pb.go
// Code generated by https://github.com/go-dev-frame/sponge, DO NOT EDIT.

package v1

import (
	"context"
	"errors"
	"strings"

	"github.com/gin-gonic/gin"
	"go.uber.org/zap"

	"github.com/go-dev-frame/sponge/pkg/errcode"
	"github.com/go-dev-frame/sponge/pkg/gin/middleware"
)

type FinancialTransactionLogicer interface {
	Create(ctx context.Context, req *CreateFinancialTransactionRequest) (*CreateFinancialTransactionReply, error)
	DeleteByID(ctx context.Context, req *DeleteFinancialTransactionByIDRequest) (*DeleteFinancialTransactionByIDReply, error)
	UpdateByID(ctx context.Context, req *UpdateFinancialTransactionByIDRequest) (*UpdateFinancialTransactionByIDReply, error)
	GetByID(ctx context.Context, req *GetFinancialTransactionByIDRequest) (*GetFinancialTransactionByIDReply, error)
	List(ctx context.Context, req *ListFinancialTransactionRequest) (*ListFinancialTransactionReply, error)
}

type FinancialTransactionOption func(*financialTransactionOptions)

type financialTransactionOptions struct {
	isFromRPC  bool
	responser  errcode.Responser
	zapLog     *zap.Logger
	httpErrors []*errcode.Error
	rpcStatus  []*errcode.RPCStatus
	wrapCtxFn  func(c *gin.Context) context.Context
}

func (o *financialTransactionOptions) apply(opts ...FinancialTransactionOption) {
	for _, opt := range opts {
		opt(o)
	}
}

func WithFinancialTransactionHTTPResponse() FinancialTransactionOption {
	return func(o *financialTransactionOptions) {
		o.isFromRPC = false
	}
}

func WithFinancialTransactionRPCResponse() FinancialTransactionOption {
	return func(o *financialTransactionOptions) {
		o.isFromRPC = true
	}
}

func WithFinancialTransactionResponser(responser errcode.Responser) FinancialTransactionOption {
	return func(o *financialTransactionOptions) {
		o.responser = responser
	}
}

func WithFinancialTransactionLogger(zapLog *zap.Logger) FinancialTransactionOption {
	return func(o *financialTransactionOptions) {
		o.zapLog = zapLog
	}
}

func WithFinancialTransactionErrorToHTTPCode(e ...*errcode.Error) FinancialTransactionOption {
	return func(o *financialTransactionOptions) {
		o.httpErrors = e
	}
}

func WithFinancialTransactionRPCStatusToHTTPCode(s ...*errcode.RPCStatus) FinancialTransactionOption {
	return func(o *financialTransactionOptions) {
		o.rpcStatus = s
	}
}

func WithFinancialTransactionWrapCtx(wrapCtxFn func(c *gin.Context) context.Context) FinancialTransactionOption {
	return func(o *financialTransactionOptions) {
		o.wrapCtxFn = wrapCtxFn
	}
}

func RegisterFinancialTransactionRouter(
	iRouter gin.IRouter,
	groupPathMiddlewares map[string][]gin.HandlerFunc,
	singlePathMiddlewares map[string][]gin.HandlerFunc,
	iLogic FinancialTransactionLogicer,
	opts ...FinancialTransactionOption) {

	o := &financialTransactionOptions{}
	o.apply(opts...)

	if o.responser == nil {
		o.responser = errcode.NewResponser(o.isFromRPC, o.httpErrors, o.rpcStatus)
	}
	if o.zapLog == nil {
		o.zapLog, _ = zap.NewProduction()
	}

	r := &financialTransactionRouter{
		iRouter:               iRouter,
		groupPathMiddlewares:  groupPathMiddlewares,
		singlePathMiddlewares: singlePathMiddlewares,
		iLogic:                iLogic,
		iResponse:             o.responser,
		zapLog:                o.zapLog,
		wrapCtxFn:             o.wrapCtxFn,
	}
	r.register()
}

type financialTransactionRouter struct {
	iRouter               gin.IRouter
	groupPathMiddlewares  map[string][]gin.HandlerFunc
	singlePathMiddlewares map[string][]gin.HandlerFunc
	iLogic                FinancialTransactionLogicer
	iResponse             errcode.Responser
	zapLog                *zap.Logger
	wrapCtxFn             func(c *gin.Context) context.Context
}

func (r *financialTransactionRouter) register() {
	r.iRouter.Handle("POST", "/api/v1/financialTransaction", r.withMiddleware("POST", "/api/v1/financialTransaction", r.Create_16)...)
	r.iRouter.Handle("DELETE", "/api/v1/financialTransaction/:id", r.withMiddleware("DELETE", "/api/v1/financialTransaction/:id", r.DeleteByID_14)...)
	r.iRouter.Handle("PUT", "/api/v1/financialTransaction/:id", r.withMiddleware("PUT", "/api/v1/financialTransaction/:id", r.UpdateByID_14)...)
	r.iRouter.Handle("GET", "/api/v1/financialTransaction/:id", r.withMiddleware("GET", "/api/v1/financialTransaction/:id", r.GetByID_14)...)
	r.iRouter.Handle("POST", "/api/v1/financialTransaction/list", r.withMiddleware("POST", "/api/v1/financialTransaction/list", r.List_16)...)

}

func (r *financialTransactionRouter) withMiddleware(method string, path string, fn gin.HandlerFunc) []gin.HandlerFunc {
	handlerFns := []gin.HandlerFunc{}

	// determine if a route group is hit or miss, left prefix rule
	for groupPath, fns := range r.groupPathMiddlewares {
		if groupPath == "" || groupPath == "/" {
			handlerFns = append(handlerFns, fns...)
			continue
		}
		size := len(groupPath)
		if len(path) < size {
			continue
		}
		if groupPath == path[:size] {
			handlerFns = append(handlerFns, fns...)
		}
	}

	// determine if a single route has been hit
	key := strings.ToUpper(method) + "->" + path
	if fns, ok := r.singlePathMiddlewares[key]; ok {
		handlerFns = append(handlerFns, fns...)
	}

	return append(handlerFns, fn)
}

func (r *financialTransactionRouter) Create_16(c *gin.Context) {
	req := &CreateFinancialTransactionRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.Create(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *financialTransactionRouter) DeleteByID_14(c *gin.Context) {
	req := &DeleteFinancialTransactionByIDRequest{}
	var err error

	if err = c.ShouldBindUri(req); err != nil {
		r.zapLog.Warn("ShouldBindUri error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	if err = c.ShouldBindQuery(req); err != nil {
		r.zapLog.Warn("ShouldBindQuery error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.DeleteByID(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *financialTransactionRouter) UpdateByID_14(c *gin.Context) {
	req := &UpdateFinancialTransactionByIDRequest{}
	var err error

	if err = c.ShouldBindUri(req); err != nil {
		r.zapLog.Warn("ShouldBindUri error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.UpdateByID(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *financialTransactionRouter) GetByID_14(c *gin.Context) {
	req := &GetFinancialTransactionByIDRequest{}
	var err error

	if err = c.ShouldBindUri(req); err != nil {
		r.zapLog.Warn("ShouldBindUri error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	if err = c.ShouldBindQuery(req); err != nil {
		r.zapLog.Warn("ShouldBindQuery error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.GetByID(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *financialTransactionRouter) List_16(c *gin.Context) {
	req := &ListFinancialTransactionRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.List(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}