# _15_appliance_store/docs/apis.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "store api docs",
    "version": "2.0"
  },
...
  "host": "localhost:8080",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/financialTransaction": {
      "post": {
        "summary": "create financialTransaction",
        "operationId": "financialTransaction_Create",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateFinancialTransactionReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateFinancialTransactionRequest"
            }
          }
        ],
        "tags": [
          "financialTransaction"
        ]
      }
    },
    "/api/v1/financialTransaction/list": {
      "post": {
        "summary": "list of financialTransaction by query parameters",
        "operationId": "financialTransaction_List",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListFinancialTransactionReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1ListFinancialTransactionRequest"
            }
          }
        ],
        "tags": [
          "financialTransaction"
        ]
      }
    },
    "/api/v1/financialTransaction/{id}": {
      "get": {
        "summary": "get financialTransaction by id",
        "operationId": "financialTransaction_GetByID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetFinancialTransactionByIDReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "uint64"
          }
        ],
        "tags": [
          "financialTransaction"
        ]
      },
      "delete": {
        "summary": "delete financialTransaction by id",
        "operationId": "financialTransaction_DeleteByID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteFinancialTransactionByIDReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "uint64"
          }
        ],
        "tags": [
          "financialTransaction"
        ]
      },
      "put": {
        "summary": "update financialTransaction by id",
        "operationId": "financialTransaction_UpdateByID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateFinancialTransactionByIDReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "uint64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "integer",
                  "format": "int32"
                },
                "amount": {
                  "type": "string"
                },
                "relatedID": {
                  "type": "string"
                },
                "transactionTime": {
                  "type": "string"
                },
                "operator": {
                  "type": "integer",
                  "format": "int64"
                },
                "remark": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": [
          "financialTransaction"
        ]
      }
    },
...
  },
  "definitions": {
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "typesColumn": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "exp": {
          "type": "string"
        },
        "value": {
          "type": "string"
        },
        "logic": {
          "type": "string"
        }
      }
    },
    "typesParams": {
      "type": "object",
      "properties": {
        "page": {
          "type": "integer",
          "format": "int32"
        },
        "limit": {
          "type": "integer",
          "format": "int32"
        },
        "sort": {
          "type": "string"
        },
        "columns": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/typesColumn"
          }
        }
      }
    },
...
    "v1CreateFinancialTransactionReply": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "uint64"
        }
      }
    },
    "v1CreateFinancialTransactionRequest": {
      "type": "object",
      "properties": {
        "type": {
          "type": "integer",
          "format": "int32"
        },
        "amount": {
          "type": "string"
        },
        "relatedID": {
          "type": "string"
        },
        "transactionTime": {
          "type": "string"
        },
        "operator": {
          "type": "integer",
          "format": "int64"
        },
        "remark": {
          "type": "string"
        }
      }
    },
...
    "v1DeleteFinancialTransactionByIDReply": {
      "type": "object"
    },
...
    "v1FinancialTransaction": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "uint64"
        },
        "type": {
          "type": "integer",
          "format": "int32"
        },
        "amount": {
          "type": "string"
        },
        "relatedID": {
          "type": "string"
        },
        "transactionTime": {
          "type": "string"
        },
        "operator": {
          "type": "integer",
          "format": "int64"
        },
        "remark": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
...
    "v1GetFinancialTransactionByIDReply": {
      "type": "object",
      "properties": {
        "financialTransaction": {
          "$ref": "#/definitions/v1FinancialTransaction"
        }
      }
    },
...
    "v1ListFinancialTransactionReply": {
      "type": "object",
      "properties": {
        "total": {
          "type": "integer",
          "format": "int64"
        },
        "financialTransactions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1FinancialTransaction"
          }
        }
      }
    },
    "v1ListFinancialTransactionRequest": {
      "type": "object",
      "properties": {
        "params": {
          "$ref": "#/definitions/typesParams"
        }
      }
    },
...
    "v1UpdateFinancialTransactionByIDReply": {
      "type": "object"
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "description": "Type Bearer your-jwt-token to Value",
      "name": "Authorization",
      "in": "header"
    }
  }
}